restructure app folder
