+++ src/app/directives/IsAuthenticated.js
