+++ src/app/schema.js
@@ -1,12 +1,14 @@
 import path from 'path';
 import { mergeTypes, mergeResolvers, fileLoader } from 'merge-graphql-schemas';
 import { makeExecutableSchema } from 'apollo-server';
-import { IsAuthenticatedDirective } from './SchemaDirectives';
+import { IsAuthenticatedDirective } from './directives';
 
-const typesArray = fileLoader(path.join(__dirname, './**/typeDef.js'));
+const typesArray = fileLoader(path.join(__dirname, './modules/**/typeDef.js'));
 const typeDefs = mergeTypes(typesArray, { all: true });
 
-const resolversArray = fileLoader(path.join(__dirname, './**/resolvers.js'));
+const resolversArray = fileLoader(
+  path.join(__dirname, './modules/**/resolvers.js'),
+);
 const resolvers = mergeResolvers(resolversArray);
 
 const schema = makeExecutableSchema({
